+++ tombio/vis1/vis1.js
@@ -18,7 +18,7 @@
         this.metadata.publisher = "Field Studies Council";
         this.metadata.location = "Shrewsbury, England";
         this.metadata.contact = "richardb@field-studies-council.org";
-        this.metadata.version = '1.0';
+        this.metadata.version = '1.1';
 
         //Other initialisations
         this.displayToolTips = true;
@@ -585,9 +585,7 @@
         params.push("expand=" + ranges.join(","));
 
         //Generate the full URL
-        var url = encodeURI(window.location.href.split('?')[0] + "?" + params.join("&"));
-        _this.copyTextToClipboard(url);
-        console.log(url);
+        _this.createViewURL(params);
     }
 
     function expandTaxon(d, i) {