+++ tombio/vis2/vis2.js
@@ -18,7 +18,7 @@
         this.metadata.publisher = "Field Studies Council";
         this.metadata.location = "Shrewsbury, England";
         this.metadata.contact = "richardb@field-studies-council.org";
-        this.metadata.version = '1.0';
+        this.metadata.version = '1.1';
 
         //Controls works with character state input controls
         this.charStateInput = true;
@@ -497,9 +497,8 @@
         params.push("imgtips=" + _this.displayToolTips);
 
         //Generate the full URL
-        var url = encodeURI(window.location.href.split('?')[0] + "?" + params.join("&"));
-        _this.copyTextToClipboard(url);
-        console.log(url);
+
+        _this.createViewURL(params);
     }
 
 })(jQuery, this.tombiovis)